+++ storybook/stories/wallets/send/WalletSend.stories.js
@@ -152,6 +152,8 @@ const confirmationAssets = assets.total.map((assetTotal) => {
     assetName: assetTotal.assetName,
     fingerprint,
     quantity: assetTotal.quantity,
+    decimals: 0,
+    recommendedDecimals: null,
     metadata: assetData
       ? assetData.metadata
       : {
@@ -185,6 +187,8 @@ const sendFormAssetData = assets.total.map((assetTotal) => {
     assetName: assetTotal.assetName,
     fingerprint,
     quantity: assetTotal.quantity,
+    decimals: 0,
+    recommendedDecimals: null,
     metadata: assetData
       ? assetData.metadata
       : {
@@ -218,6 +222,7 @@ storiesOf('Wallets|Send', module)
       hasAssets={boolean('hasAssets', false)}
       selectedAsset={null}
       onUnsetActiveAssetFingerprint={() => {}}
+      isAddressFromSameWallet={boolean('isAddressFromSameWallet', false)}
     />
   ))
   .add('Send - Hardware wallet verifying transaction', () => (
@@ -241,6 +246,7 @@ storiesOf('Wallets|Send', module)
       hasAssets={boolean('hasAssets', false)}
       selectedAsset={null}
       onUnsetActiveAssetFingerprint={() => {}}
+      isAddressFromSameWallet={boolean('isAddressFromSameWallet', false)}
     />
   ))
   .add('Send - Hardware wallet verifying transaction succeeded', () => (
@@ -264,6 +270,7 @@ storiesOf('Wallets|Send', module)
       hasAssets={boolean('hasAssets', false)}
       selectedAsset={null}
       onUnsetActiveAssetFingerprint={() => {}}
+      isAddressFromSameWallet={boolean('isAddressFromSameWallet', false)}
     />
   ))
   .add('Send - Hardware wallet verifying transaction failed', () => (
@@ -287,6 +294,7 @@ storiesOf('Wallets|Send', module)
       hasAssets={boolean('hasAssets', false)}
       selectedAsset={null}
       onUnsetActiveAssetFingerprint={() => {}}
+      isAddressFromSameWallet={boolean('isAddressFromSameWallet', false)}
     />
   ))
   .add('Send - With Assets', () => (
@@ -313,6 +321,7 @@ storiesOf('Wallets|Send', module)
       hasAssets={boolean('hasAssets', true)}
       selectedAsset={null}
       onUnsetActiveAssetFingerprint={() => {}}
+      isAddressFromSameWallet={boolean('isAddressFromSameWallet', false)}
     />
   ))
   .add('Wallet Send Confirmation Dialog With Assets', () => (