+++ source/renderer/app/api/api.js
@@ -117,6 +117,7 @@ import {
   WALLET_RECOVERY_PHRASE_WORD_COUNT,
 } from '../config/cryptoConfig';
 import { currencyConfig } from '../config/currencyConfig';
+import { ASSETS_PREDEFINED_DECIMALS } from '../config/assetsConfig';
 
 // Addresses Types
 import type {
@@ -236,6 +237,7 @@ import type {
   GetAssetsResponse,
   ApiAsset,
 } from './assets/types';
+import type { AssetLocalData } from './utils/localStorage';
 import Asset from '../domains/Asset';
 import { getAssets } from './assets/requests/getAssets';
 import { getAccountPublicKey } from './wallets/requests/getAccountPublicKey';
@@ -636,7 +638,16 @@ export default class AdaApi {
       logger.debug('AdaApi::getAssets success', {
         assets: response,
       });
-      const assets = response.map((asset) => _createAssetFromServerData(asset));
+      const assetsLocaldata = await global.daedalus.api.localStorage.getAssetsLocalData();
+      logger.debug('AdaApi::getAssetsLocalData success', {
+        assetsLocaldata,
+      });
+      const assets = response.map((asset) =>
+        _createAssetFromServerData(
+          asset,
+          assetsLocaldata[asset.policy_id + asset.asset_name] || {}
+        )
+      );
       return new Promise((resolve) =>
         resolve({ assets, total: response.length })
       );
@@ -2833,18 +2844,23 @@ const _createTransactionFromServerData = action(
 
 const _createAssetFromServerData = action(
   'AdaApi::_createAssetFromServerData',
-  (data: ApiAsset) => {
+  (data: ApiAsset, localData: AssetLocalData) => {
     const {
       policy_id: policyId,
       asset_name: assetName,
       fingerprint,
       metadata,
     } = data;
+    const { decimals } = localData;
+    const recommendedDecimals =
+      ASSETS_PREDEFINED_DECIMALS[policyId + assetName];
     return new Asset({
       policyId,
       assetName,
       fingerprint,
       metadata,
+      decimals,
+      recommendedDecimals,
     });
   }
 );