+++ source/renderer/app/stores/AssetsStore.js
@@ -4,7 +4,11 @@ import { get } from 'lodash';
 import Store from './lib/Store';
 import Request from './lib/LocalizedRequest';
 import Asset from '../domains/Asset';
-import type { GetAssetsResponse } from '../api/assets/types';
+import { requestGetter } from '../utils/storesUtils';
+import type {
+  GetAssetsResponse,
+  WalletSummaryAsset,
+} from '../api/assets/types';
 
 type WalletId = string;
 
@@ -12,14 +16,24 @@ export default class AssetsStore extends Store {
   ASSETS_REFRESH_INTERVAL: number = 1 * 60 * 1000; // 1 minute | unit: milliseconds
 
   @observable activeAssetFingerprint: ?string = null;
-
+  @observable editingsAsset: ?WalletSummaryAsset = null;
   @observable assetsRequests: {
     [key: WalletId]: Request<GetAssetsResponse>,
   } = {};
 
+  // REQUESTS
+  @observable
+  getAssetSettingsDialogWasOpenedRequest: Request<void> = new Request(
+    this.api.localStorage.getAssetSettingsDialogWasOpened
+  );
+
   setup() {
     setInterval(this._refreshAssetsData, this.ASSETS_REFRESH_INTERVAL);
-    const { wallets: walletsActions } = this.actions;
+    const { assets: assetsActions, wallets: walletsActions } = this.actions;
+    assetsActions.onAssetSettingsOpen.listen(this._onAssetSettingsOpen);
+    assetsActions.onAssetSettingsSubmit.listen(this._onAssetSettingsSubmit);
+    assetsActions.onAssetSettingsCancel.listen(this._onAssetSettingsCancel);
+
     walletsActions.refreshWalletsDataSuccess.once(this._refreshAssetsData);
     walletsActions.setActiveAssetFingerprint.listen(
       this._setActiveAssetFingerprint
@@ -50,7 +64,42 @@ export default class AssetsStore extends Store {
   getAssetDetails = (policyId: string, assetName: string): ?Asset =>
     this.details[policyId + assetName];
 
+  @computed get assetSettingsDialogWasOpened(): boolean {
+    return requestGetter(this.getAssetSettingsDialogWasOpenedRequest, false);
+  }
+
   // =================== PRIVATE ==================
+
+  @action _onAssetSettingsOpen = ({ asset }: { asset: WalletSummaryAsset }) => {
+    this.editingsAsset = asset;
+    this.api.localStorage.setAssetSettingsDialogWasOpened();
+    this.getAssetSettingsDialogWasOpenedRequest.execute();
+  };
+
+  @action _onAssetSettingsSubmit = async ({
+    asset,
+    decimals,
+  }: {
+    asset: WalletSummaryAsset,
+    decimals: number,
+  }) => {
+    this.editingsAsset = null;
+    const { policyId, assetName } = asset;
+    const assetDomain = this.getAssetDetails(policyId, assetName);
+    if (assetDomain) {
+      assetDomain.update({
+        decimals,
+      });
+    }
+    this._refreshAssetsData();
+    await this.api.localStorage.setAssetLocalData(policyId, assetName, {
+      decimals,
+    });
+  };
+
+  @action _onAssetSettingsCancel = () => {
+    this.editingsAsset = null;
+  };
 
   @action _refreshAssetsData = () => {
     if (this.stores.networkStatus.isConnected) {