+++ storybook/stories/settings/general/General.stories.js
@@ -4,6 +4,7 @@ import { findKey } from 'lodash';
 import { boolean, number } from '@storybook/addon-knobs';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
+import { withState } from '@dump247/storybook-state';
 import SettingsWrapper from '../utils/SettingsWrapper';
 import {
   DATE_ENGLISH_OPTIONS,
@@ -21,6 +22,10 @@ import DisplaySettings from '../../../../source/renderer/app/components/settings
 import SupportSettings from '../../../../source/renderer/app/components/settings/categories/SupportSettings';
 import TermsOfUseSettings from '../../../../source/renderer/app/components/settings/categories/TermsOfUseSettings';
 import WalletsSettings from '../../../../source/renderer/app/components/settings/categories/WalletsSettings';
+
+// Assets and helpers
+import currenciesList from '../../../../source/renderer/app/config/currenciesList.json';
+import { getLocalizedCurrenciesList } from '../../../../source/renderer/app/config/currencyConfig';
 
 const getParamName = (obj, itemName): any =>
   Object.entries(obj).find((entry: [any, any]) => itemName === entry[1]);
@@ -49,21 +54,32 @@ storiesOf('Settings|General', module)
       currentTimeFormat={TIME_OPTIONS[0].value}
     />
   ))
-  .add('Wallets', () => (
-    <WalletsSettings
-      currencySelected={{
-        id: 'uniswap-state-dollar',
-        code: 'usd',
-        name: 'unified Stable Dollar',
-      }}
-      currencyRate={0.321}
-      currencyList={[]}
-      currencyIsActive
-      onSelectCurrency={action('onSelectCurrency')}
-      onToggleCurrencyIsActive={action('onToggleCurrencyIsActive')}
-      onOpenExternalLink={action('onOpenExternalLink')}
-    />
-  ))
+  .add(
+    'Wallets',
+    withState(
+      {
+        currencySelected: {
+          id: 'uniswap-state-dollar',
+          code: 'usd',
+          name: 'unified Stable Dollar',
+        },
+      },
+      (store) => (
+        <WalletsSettings
+          currencySelected={store.state.currencySelected}
+          currencyRate={0.321}
+          currencyList={getLocalizedCurrenciesList(currenciesList, 'en-US')}
+          currencyIsActive
+          onSelectCurrency={(code) =>
+            store.set({ currencySelected: currenciesList[code] })
+          }
+          onToggleCurrencyIsActive={action('onToggleCurrencyIsActive')}
+          onOpenExternalLink={action('onOpenExternalLink')}
+          hasSearch={boolean('hasSearch', true)}
+        />
+      )
+    )
+  )
   .add('Stake Pools', () => (
     <StakePoolsSettings
       onSelectSmashServerUrl={action('onSelectSmashServerUrl')}