+++ storybook/stories/wallets/summary/WalletSummary.stories.js
@@ -12,7 +12,7 @@ import {
   generateWallet,
 } from '../../_support/utils';
 import WalletsWrapper from '../_utils/WalletsWrapper';
-import currencyList from '../_utils/currencies.json';
+import currenciesList from '../../../../source/renderer/app/config/currenciesList.json';
 
 // Screens
 import WalletSummary from '../../../../source/renderer/app/components/wallet/summary/WalletSummary';
@@ -136,6 +136,8 @@ const walletAssets = assets.total.map((assetTotal) => {
     assetName: assetTotal.assetName,
     fingerprint,
     quantity: assetTotal.quantity,
+    decimals: 0,
+    recommendedDecimals: null,
     metadata: assetData
       ? assetData.metadata
       : {
@@ -173,18 +175,11 @@ storiesOf('Wallets|Summary', module)
       currencyIsActive = false;
     }
 
-    const currencySelected = select(
-      'currencySelected',
-      currencyList.reduce((obj, currency) => {
-        obj[`${currency.id} - ${currency.name}`] = currency;
-        return obj;
-      }, {}),
-      {
-        id: 'uniswap-state-dollar',
-        symbol: 'usd',
-        name: 'unified Stable Dollar',
-      }
-    );
+    const currencySelected = select('currencySelected', currenciesList, {
+      id: 'uniswap-state-dollar',
+      symbol: 'usd',
+      name: 'unified Stable Dollar',
+    });
 
     return (
       <WalletSummary
@@ -207,7 +202,12 @@ storiesOf('Wallets|Summary', module)
         isLoadingAssets={boolean('isLoadingAssets', false)}
         onOpenAssetSend={action('onOpenAssetSend')}
         onCopyAssetItem={action('onCopyAsset')}
+        onAssetSettings={action('onAssetSettings')}
         hasAssetsEnabled={boolean('hasAssetsEnabled', true)}
+        assetSettingsDialogWasOpened={boolean(
+          'assetSettingsDialogWasOpened',
+          false
+        )}
         onExternalLinkClick={action('onExternalLinkClick')}
       />
     );